Moved the non-features back to entity
