+++ apps/verification-portal/src/lib/features/web3InboxNotifications.ts
@@ -1,5 +1,5 @@
 import type { NotifyClientTypes } from '@walletconnect/notify-client';
-import { formatDistanceToNowStrict } from 'date-fns';
+
 import { markNotificationsAsRead } from '$lib/features/web3Inbox';
 
 export let isPending = false;
@@ -16,23 +16,4 @@ export async function markAsRead(
       isPending = false;
       notification.isRead = true;
     });
-}
-
-export function findNotificationTypeByTopic(
-  topic: string | undefined,
-  notificationTypes: NotifyClientTypes.ScopeMap[]
-) {
-  if (!notificationTypes || topic === undefined) return;
-  const type = notificationTypes.find((type) => String(type.id) === topic);
-  return type ? type : undefined;
-}
-
-export function formatDate(sentAt: number) {
-  try {
-    return formatDistanceToNowStrict(sentAt, {
-      addSuffix: true,
-    });
-  } catch (error) {
-    console.error(error);
-  }
 }
